+++ packages/zen-persistent/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sylphx/zen-persistent
 
+## 7.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @sylphx/zen@3.3.0
+  - @sylphx/zen-patterns@4.0.0
+
 ## 6.0.0
 
 ### Patch Changes