+++ packages/zen-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sylphx/zen-react
 
+## 3.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @sylphx/zen@3.3.0
+
 ## 3.0.3
 
 ### Patch Changes